implement storage state flush to l0

+++ src/state/storage_state_options.rs
@@ -3,6 +3,6 @@ use std::path::PathBuf;
 pub struct StorageStateOptions {
     pub sst_max_size_bytes: usize,
     pub block_max_size_bytes: usize,
-    pub block_cache_size_bytes: usize,
+    pub block_cache_size_bytes: u64,
     pub path: PathBuf,
 }
